% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableDataSourceDescriptions.R
\name{availableDataSourceDescriptions}
\alias{availableDataSourceDescriptions}
\title{Get available data source descriptions}
\usage{
availableDataSourceDescriptions()
}
\value{
A character vector with names of available data sources.
}
\description{
Provides a list of all available data source descriptions stored in the
oxford-pharmacoepi/DataSourceDescriptionsLibrary GitHub repository.
}
\examples{
\dontrun{
library(dataSourceDescriptions)

availableDataSourceDescriptions()
}
}
