% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSourceDescriptions-package.R
\docType{package}
\name{DataSourceDescriptions-package}
\alias{DataSourceDescriptions}
\alias{DataSourceDescriptions-package}
\title{DataSourceDescriptions package}
\description{
Utilities for creating, validating, importing, and exporting structured OMOP
data source descriptions.
}
\author{
\strong{Maintainer}: Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})

Authors:
\itemize{
  \item Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})
}

}
\keyword{internal}
