% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDataSourceDescription.R
\name{emptyDataSourceDescription}
\alias{emptyDataSourceDescription}
\title{Create an empty \code{data_source_description} object}
\usage{
emptyDataSourceDescription()
}
\value{
An empty \code{data_source_description} object.
}
\description{
Create an empty \code{data_source_description} object
}
\examples{
emptyDataSourceDescription()

}
