% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDataSourceDescription.R
\name{exportDataSourceDescription}
\alias{exportDataSourceDescription}
\title{Export data source descriptions}
\usage{
exportDataSourceDescription(x, path, type = "json")
}
\arguments{
\item{x}{A data source description.}

\item{path}{Directory where files will be written.}

\item{type}{File type to export. Supported values are \code{"json"} and \code{"csv"}.}
}
\value{
Invisibly returns the exported file paths.
}
\description{
Export data source descriptions
}
