% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataSourceDescription.R
\name{importDataSourceDescription}
\alias{importDataSourceDescription}
\title{Import data source descriptions}
\usage{
importDataSourceDescription(path, type = NULL, recursive = FALSE)
}
\arguments{
\item{path}{A file path or directory path. When a directory is supplied,
matching files are imported from the directory.}

\item{type}{File type to import. Supported values are \code{"json"} and \code{"csv"}.
If \code{NULL}, both supported file types are detected from file extensions.}

\item{recursive}{Whether to search directories recursively.}
}
\value{
A \code{data_source_description} object.
}
\description{
Import data source descriptions
}
