% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDataSourceDescription.R
\name{newDataSourceDescription}
\alias{newDataSourceDescription}
\title{\code{data_source_description} object constructor}
\usage{
newDataSourceDescription(x)
}
\arguments{
\item{x}{A data source description, or a named list of data source
descriptions.}
}
\value{
A \code{data_source_description} object.
}
\description{
\code{data_source_description} object constructor
}
