% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{visualiseDataSourceDescription}
\alias{visualiseDataSourceDescription}
\title{Visualise bundled data source descriptions}
\usage{
visualiseDataSourceDescription(dataSourceName = NULL)
}
\arguments{
\item{dataSourceName}{Character vector of data source acronyms to visualise.
If \code{NULL}, all bundled data source descriptions are shown.}
}
\value{
A \code{bslib} tabset.
}
\description{
Creates an HTML view of one or more bundled data source descriptions. The
returned object is a \code{bslib} tabset that displays in the RStudio Viewer when
printed interactively.
}
\examples{
\dontrun{
visualiseDataSourceDescription("CPRD GOLD")
}

}
