% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDataSourceDescription.R
\name{downloadDataSourceDescription}
\alias{downloadDataSourceDescription}
\title{Download data source descriptions}
\usage{
downloadDataSourceDescription(names, path)
}
\arguments{
\item{names}{A character vector of data source names to download. You can use
\code{availableDataSourceDescriptions()} to find available descriptions. If
NULL, all descriptions will be saved}

\item{path}{A character string specifying the local directory where the
files should be saved.}
}
\value{
Invisible \code{NULL}. Called for its side effect of downloading files.
}
\description{
Downloads specific data source descriptions from the
oxford-pharmacoepi/DataSourceDescriptionsLibrary GitHub repository and saves
them to a local directory.
}
\examples{
\dontrun{
# Assuming you have a folder named "data_sources" in your project directory
downloadDataSourceDescription(
  names = c("CPRD_AURUM"),
  path = here::here("data_sources")
)
}
}
